Fix Windows build: remove IMG_Init (removed in SDL3_image), fix C4267 warnings
- Remove IMG_Init/IMG_INIT_JPG/IMG_INIT_PNG/IMG_GetError block in game.cpp:
  SDL3_image loads formats on demand, no explicit Init required
- Fix C4267 (size_t->int) in sector.h, panel.h, game.cpp, gamestate.cpp,
  image.cpp, panel.cpp, sector.cpp by adding explicit (int) casts
- Remove stale /usr/X11R6/lib/ from Makefile LINKPATH (macOS linker warning)

+++ gamestate.cpp
@@ -1623,7 +1623,7 @@ void PlayingGameState::draw() {
 	//Vector soldier_list(n_players_c * 250);
 	int n_total_soldiers = 0;
 	for(int i=0;i<n_players_c;i++) {
-		n_total_soldiers += soldiers[i].size();
+		n_total_soldiers += (int)soldiers[i].size();
 	}
 	Soldier **soldier_list = new Soldier *[n_total_soldiers];
 	for(int i=0,c=0;i<n_players_c;i++) {
@@ -1661,14 +1661,14 @@ void PlayingGameState::draw() {
 		}
 	}
 
-	for(int i=effects.size()-1;i>=0;i--) {
+	for(int i=(int)effects.size()-1;i>=0;i--) {
 		TimedEffect *effect = effects.at(i);
 		if( effect->render() ) {
 			effects.erase(effects.begin() + i);
 			delete effect;
 		}
 	}
-	for(int i=ammo_effects.size()-1;i>=0;i--) {
+	for(int i=(int)ammo_effects.size()-1;i>=0;i--) {
 		TimedEffect *effect = ammo_effects.at(i);
 		if( effect->render() ) {
 			ammo_effects.erase(ammo_effects.begin() + i);
